requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Deep Learning & Generative Models
torch>=2.3.0
sdv>=1.29.0                   # For CTGAN baseline
ctgan>=0.11.0                 # CTGAN implementation

# LLM Integration
litellm>=1.0.0                # Universal LLM interface
openai>=1.0.0                 # OpenAI API
anthropic>=0.18.0             # Anthropic Claude API
google-generativeai>=0.3.0    # Google Gemini API

# Multi-Agent Framework
pyautogen>=0.2.0              # Microsoft AutoGen

# Fairness & Bias
fairlearn>=0.9.0              # Fairness metrics
aif360>=0.5.0                 # AI Fairness 360

# Machine Learning
xgboost>=2.0.0                # Gradient boosting

# Distance Metrics
gower>=0.1.2                  # Gower distance for mixed types

# Clustering & Dimensionality Reduction
scikit-learn>=1.3.0           # Includes DBSCAN, t-SNE, PCA

# Utilities
python-dotenv>=1.0.0          # Environment variable management
pyyaml>=6.0                   # Config file parsing
tqdm>=4.65.0                  # Progress bars
joblib>=1.3.0                 # Parallel processing

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Optional: AutoGen Studio
autogenstudio>=0.0.1          # Visual debugging tool (optional)

